backend/requirements.txt
# =============================================================================
# Ninai Backend - Python Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Web Framework
# -----------------------------------------------------------------------------
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3
psycopg2-binary==2.9.9

# -----------------------------------------------------------------------------
# Validation & Serialization
# -----------------------------------------------------------------------------
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# -----------------------------------------------------------------------------
# Authentication & Security
# -----------------------------------------------------------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pyotp==2.9.0
webauthn==2.1.0
qrcode==7.4.2

# -----------------------------------------------------------------------------
# Caching
# -----------------------------------------------------------------------------
redis==5.0.1
hiredis==2.3.2

# -----------------------------------------------------------------------------
# Vector Database
# -----------------------------------------------------------------------------
qdrant-client==1.7.0

# -----------------------------------------------------------------------------
# Background Tasks
# -----------------------------------------------------------------------------
celery==5.3.6
flower==2.0.1

# -----------------------------------------------------------------------------
# Elasticsearch
# -----------------------------------------------------------------------------
elasticsearch[async]==8.11.1

# -----------------------------------------------------------------------------
# LLM & Embeddings (Optional)
# -----------------------------------------------------------------------------
openai==1.9.0
anthropic==0.12.0
tiktoken==0.5.2

# -----------------------------------------------------------------------------
# HTTP Client
# -----------------------------------------------------------------------------
httpx==0.26.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dateutil==2.8.2
orjson==3.9.12
structlog==24.1.0
tenacity==8.2.3
pypdf==4.0.1
pillow==10.2.0
pytesseract==0.3.10
python-docx==1.1.0

# -----------------------------------------------------------------------------
# Development & Testing
# -----------------------------------------------------------------------------
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0
faker==22.2.0

# -----------------------------------------------------------------------------
# Code Quality
# -----------------------------------------------------------------------------
black==24.1.0
isort==5.13.2
ruff==0.1.13
mypy==1.8.0
